+++ test/e2e/kubectl.go
@@ -45,6 +45,7 @@ const (
 )
 
 var _ = Describe("kubectl", func() {
+	defer GinkgoRecover()
 	var c *client.Client
 	var ns string
 	var testingNs *api.Namespace
@@ -65,11 +66,12 @@ var _ = Describe("kubectl", func() {
 	})
 
 	Describe("update-demo", func() {
-		var (
+		var updateDemoRoot, nautilusPath, kittenPath string
+		BeforeEach(func() {
 			updateDemoRoot = filepath.Join(testContext.RepoRoot, "examples/update-demo")
-			nautilusPath   = filepath.Join(updateDemoRoot, "nautilus-rc.yaml")
-			kittenPath     = filepath.Join(updateDemoRoot, "kitten-rc.yaml")
-		)
+			nautilusPath = filepath.Join(updateDemoRoot, "nautilus-rc.yaml")
+			kittenPath = filepath.Join(updateDemoRoot, "kitten-rc.yaml")
+		})
 
 		It("should create and stop a replication controller", func() {
 			defer cleanup(nautilusPath, ns, updateDemoSelector)
@@ -105,7 +107,10 @@ var _ = Describe("kubectl", func() {
 	})
 
 	Describe("guestbook", func() {
-		var guestbookPath = filepath.Join(testContext.RepoRoot, "examples/guestbook")
+		var guestbookPath string
+		BeforeEach(func() {
+			guestbookPath = filepath.Join(testContext.RepoRoot, "examples/guestbook")
+		})
 
 		It("should create and stop a working application", func() {
 			if !providerIs("gce", "gke") {
